Add Projects module pattern
This module pattern acts as the hub of projects, every project that
gets created by the factory function will be added to the projectsObject
which will then be added to the localStorage, each project will be represented by a key
and each key will hold and array of items of each project.

There're functions to add and remove projects from the projects object and functions
to update and retrieve projects data from local storage. The update function will be used
everytime there's a change in the projects object and the retrieve function will be triggered
everytime there's a page reload.