Write renderProjects and generateProject functions
generateProject generates a project card based in the
project name passed to it, which is a project key name that's
included in the projects object. renderProjects retrieves the projects object
stored in local storage, parses it, stores it in a variable and then iterates
over it generatating project cards with the generateProject function.